emailer: config: increase task_compose_emails_loop_sleep value
The current code is unable to detect some duplicated events if an email
has been sent already.

Try mitigating this duplicated email about a similar event with a longer
caching phase before email compositing.

The value 400 seconds is 6.5 minutes, which seems to be enough to catch most
of the detected ocurrences so far.

Bug: T286135
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I05e1027e91f7909c57173863b43494f04881abf6