emailer: introduce k8s liveness probe support
This patch introduces a small HTTP listener for kubernetes liveness probes.

The timeout for task_watch_pod() should be lowered, because is the main
blocking point of the app, and it can make the HTTP server miss probes. At this
point, I see no value in it being a config option, so just hardcode it.

By having the HTTP server iterate on the same event loop as the rest of the
daemon (as oppose to in a different worker thread), we make sure the liveness
probe is effectively checking that the event loop is iterating. This should
help discover cases in which the event loop has blocked for whatever reason
(which is something we've seen before).

Bug: T320405
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I343d8c055bcffeb2ae201027df36e51ddfdc5093